+++ src/main/kotlin/com/coder/toolbox/sdk/v2/models/Template.kt
@@ -8,4 +8,7 @@ import java.util.UUID
 data class Template(
     @Json(name = "id") val id: UUID,
     @Json(name = "active_version_id") val activeVersionID: UUID,
+    @Json(name = "name") val name: String,
+    @Json(name = "display_name") val displayName: String,
+    @Json(name = "deprecated") val deprecated: Boolean,
 )